edit for comments and comments/tags css

+++ frontend/components/tag/tag_index_item.jsx
@@ -2,25 +2,32 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 
 const TagIndexItem = props => {
-  const delTag = props.userId === props.photo.user_id ?
-  <Link
-    className="tag-index-item-delete"
-    onClick={() => props.deleteTag(props.tag.id, props.photo.id)}
-    to={`/photos/${props.photo.id}`}
-  >x</Link> : ""
-
-  return (
-    <li>
-      <div className="tag-index-item-container">
-        <div className="tag-index-item">
-          <Link to={`/search/photos/${props.tag.word}`}>
+  if (props.userId === props.photo.user_id) {
+    return(
+      <li>
+        <div className="tag-index-item-photo-owner">
+          <Link className="tag-index-item-word" to={`/search/photos/${props.tag.word}`}>
             {props.tag.word}
           </Link>
-          {delTag}
+          <Link
+            className="tag-index-item-delete"
+            onClick={() => props.deleteTag(props.tag.id, props.photo.id)}
+            to={`/photos/${props.photo.id}`}
+          >x</Link>
         </div>
-      </div>
-    </li>
-  );
+      </li>
+    )
+  } else {
+    return (
+      <li>
+        <div className="tag-index-item">
+          <Link className="tag-index-item-word" to={`/search/photos/${props.tag.word}`}>
+            {props.tag.word}
+          </Link>
+        </div>
+      </li>
+    );
+  }
 }
 
 export default TagIndexItem;
